Reactor inventory app
