+++ finish/src/main/java/io/openliberty/guides/inventory/InventoryManager.java
@@ -16,28 +16,26 @@ package io.openliberty.guides.inventory;
 
 import java.io.IOException;
 import java.util.Properties;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 import javax.enterprise.context.ApplicationScoped;
 import org.eclipse.microprofile.faulttolerance.Fallback;
-import io.openliberty.guides.inventory.model.InventoryList;
+import io.openliberty.guides.inventory.model.*;
 
 @ApplicationScoped
 public class InventoryManager {
 
-  private InventoryList invList = new InventoryList();
+  private List<SystemData> systems = Collections.synchronizedList(new ArrayList<>());
   private InventoryUtils invUtils = new InventoryUtils();
 
   @Fallback(fallbackMethod = "fallbackForGet")
   public Properties get(String hostname) throws IOException {
-    Properties properties = invUtils.getProperties(hostname);
-
-    if (properties != null) {
-      invList.addToInventoryList(hostname, properties);
-    }
-    return properties;
+    return invUtils.getProperties(hostname);
   }
 
   public Properties fallbackForGet(String hostname) {
-    Properties properties = invList.findHost(hostname);
+    Properties properties = findHost(hostname);
     if (properties == null) {
       Properties msgProp = new Properties();
       msgProp.setProperty(hostname, "System is not found in the inventory");
@@ -46,8 +44,28 @@ public class InventoryManager {
     return properties;
   }
 
+  public void add(String hostname, Properties systemProps) {
+    Properties props = new Properties();
+    props.setProperty("os.name", systemProps.getProperty("os.name"));
+    props.setProperty("user.name", systemProps.getProperty("user.name"));
+ 
+    SystemData system = new SystemData(hostname, props);
+    if (!systems.contains(system)) {
+      systems.add(system);
+    }
+  }
+
   public InventoryList list() {
-    return invList;
+    return new InventoryList(systems);
+  }
+
+  private Properties findHost(String hostname) {
+    for (SystemData system : systems) {
+      if (system.getHostname().equals(hostname)) {
+        return system.getProperties();
+      }
+    }
+    return null;
   }
 }
 // end::add_fallback[]
